upload/requirements.txt
cloudevents==1.10.1
dapr==1.11.0
fastapi==0.111.0
grpcio==1.62.1
pydantic==2.4.2
requests==2.32.0
uvicorn==0.23.2
aiohttp==3.10.2